refactor(ffmpeg/transcode): relocate audio recode helpers to sibling file
Moves 4 audio-encoding-during-recode helpers from video_convert.rs
to a new audio_recode_pipeline.rs sibling:
- build_audio_transcode_filter (~94 LOC)
- drain_audio_transcode_filtered (~38 LOC)
- drain_audio_filter_to_encoder (~22 LOC)
- drain_audio_encoder_packets (~25 LOC)

Fixes asymmetry: video_pipeline.rs holds video-side recode helpers;
audio-side helpers now live in audio_recode_pipeline.rs (matching
the audio_pipeline.rs / audio_pipeline_direct.rs naming pattern).

video_convert.rs: 928 -> 745 LOC. Behavior unchanged; tests pass.

Visibility lifted from fn to pub(super) on the 4 moved helpers
(minimum viable for cross-module sibling access from
video_convert.rs's Self:: callsites).

The #[allow(clippy::too_many_arguments)] on drain_audio_transcode_filtered
moves with the function; resolving the 9-arg violation is out of scope.
